Add files via upload
Successfully got create note functionality working. When you type in a student id and employee id in the create note page, the IDs will be compared with the respective tables in database to check if they exist, so when you create a note you need to use a student id and employee id that already exist in the database. Valid student ids are: 11, 22, 33. Valid employee ids are 99, 88, 77. After signing in the noteListGUI page shows the noteIDs that exist in the database. Each of those noteIDs lead to a studentNote in the database. I intentionally disabled the remove note button, we can change that to mark as complete rather than delete. Or perhaps we can keep the delete note functionality in addition to the mark note as complete functionality.

Things to do: 
-Add button to main menu that allows you to view notes.
-Change the display box at main menu to include student name in addition instead of just noteID (You will see what I mean when you run program)
-Get the "go back" and "clear" buttons working
-Clean up the UI a bit 
-I didn't really change anything about the NoteClass.py file, don't even know if we need that
-Probably more things I haven't caught yet